requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow
opencv-python
pydantic
python-dotenv
streamlit